+++ src/overlays/actors/ovl_Boss_Tw/z_boss_tw.c
@@ -3114,7 +3114,7 @@ void BossTw_TwinrovaUpdate(Actor* thisx, PlayState* play2) {
             }
 
             if (!(info->toucher.dmgFlags & DMG_HOOKSHOT)) {
-                if (((s8)this->actor.colChkInfo.health < 30) && !swordDamage) {
+                if (((s16)this->actor.colChkInfo.health < 30) && !swordDamage) {
                     damage = 0;
                 }
 
@@ -5146,7 +5146,7 @@ void BossTw_TwinrovaChargeBlast(BossTw* this, PlayState* play) {
     Math_ApproachS(&this->actor.shape.rot.y, this->actor.yawTowardsPlayer, 5, 0x1000);
 
     if (Animation_OnFrame(&this->skelAnime, this->workf[ANIM_SW_TGT])) {
-        if ((s8)this->actor.colChkInfo.health < 100) {
+        if ((s16)this->actor.colChkInfo.health < 100) {
             sTwinrovaBlastType = Rand_ZeroFloat(1.99f);
         } else {
             if (++sFixedBlatSeq >= 4) {
@@ -5260,11 +5260,11 @@ void BossTw_TwinrovaDamage(BossTw* this, PlayState* play, u16 damage) {
         this->workf[ANIM_SW_TGT] = Animation_GetLastFrame(&gTwinrovaDamageAnim);
         this->csState1 = 1;
 
-        if ((s8)(this->actor.colChkInfo.health -= damage) < 0) {
+        if ((s16)(this->actor.colChkInfo.health -= damage) < 0) {
             this->actor.colChkInfo.health = 0;
         }
 
-        if ((s8)this->actor.colChkInfo.health <= 0) {
+        if ((s16)this->actor.colChkInfo.health <= 0) {
             BossTw_TwinrovaSetupDeathCS(this, play);
             Enemy_StartFinishingBlow(play, &this->actor);
             Actor_PlaySfx(&this->actor, NA_SE_EN_TWINROBA_YOUNG_DEAD);
